refactor: move object registry to plexe/core
