+++ plexe/core/object_registry.py
